requirements.txt
PyYAML==6.0.2
qrcode==8.0
